Add log in scenario
Paired with Ravi and Joe to create a scenario for logging into the portal

The new scenario logs in using the USERNAME and PASSWORD environment
variables which you should set before running the test.
```
export USERNAME=test
export PASSWORD=password
```
These variables should match an existing user in the test Portal.

Once the user is logged we check that the portal application home page
is displayed.

Co-authored-by: Ravi Penumarthy <ravi.penumarthy@justice.gov.uk>
Co-authored-by: Joe Giffin <joe.giffin@justice.gov.uk>