After hook for logging out of portal
There are a maximum number of sessions allowed for a particular user on
the portal, if we exceed this number then the user will not be allowed
to access the portal again.  We can clear the session with the portal by
logging out.  This change adds an after hook so that the log out page is
always called.