Log into CWA
Add feature to log into CWA, we reuse the existing steps to log into the
portal and view the portal homepage, then we click on the CWA link.

We had to add a page.has_selector? query to wait for the page to load,
otherwise the next step would fail as the CWA window hadn't loaded yet.

Once the page is loaded we switch to that window and check that the page
has the required content.

Co-authored-by: Ravi Penumarthy <ravi.penumarthy@justice.gov.uk>