Add an outcome to a submission
This feature adds an outcome sucessfully to a submission.

At certain points the elements refresh to change the layout of the page,
at these points we have added in page.find_by_id to cause capybara to
wait, until the element has loaded.  When the time to render the page
has exceeded the default wait (2 seconds) we have entered a custom wait
time.

Co-authored-by: Ravi Penumarthy <ravi.penumarthy@justice.gov.uk>