Refactor outcome page fill into module
We want to test the unhappy path and other senarios, this would involve
a lot of duplicate code so we have refactored the add outcome step to
another method.  Then we simply pass in a hash of values from the step,
this way our step file is much cleaner and duplicate code is reduced.

Co-authored-by: Ravi Penumarthy <ravi.penumarthy@justice.gov.uk>