PA and AP validation feature
Procurement area and access point validation feature added.
We use Cucumber senario outlines to pass in different values for the
case start date, procurement area and access points.  We are passing in
the case_id currently as we do not have any tear down yet which would
delete the created outcomes. This is needed as we would otherwise get
duplicate errors.  This will be fixed in future PRs.

Co-authored-by: Ravi Penumarthy <ravi.penumarthy@justice.gov.uk>