Remove ssl certificate checking from webdriver
The UAT environment has a self signed certificate, which causes a
browser error when running the tests.  This change removes the ssl
checking (there is no option to add exceptions that I could find) to
allow the tests to run.

Also updated the cwa_env.yml file with the correct values for UAT.