Delete added outcomes after a senario
Outcomes were being created but not deleted, meaning that the tests
would only run once sucessfully and then fail.  By deleting all of the
outcomes after the relevant senarios this ensures the test can be re
run.

To delete outcomes after a senario we include the @delete_outcome_after
tag to the start of the senario.

I did try to have the delete_all_outcomes method pick up from the end of
the previous senario, but couldn't get it to work in the same page
window. With more time this is something that could be looked into.  The
work around as it is currently takes a little longer to run as it logs
into the portal, navigates to the submission list, and then deletes the
outcomes.  But by doing it this way it does mean that it can be run from
any point in the tests and is not dependant on the current state.