Define `AddOutcomePage` object in a dynamic way
This page is, by definition, dinamically populated with fields that
can have different names and element IDs. This changes based on
the context value set for the Category of Law chosen.

The solution in this commit allows to define as many available
input/select elements as possible for all screens and find them by
looking at the label name to their left. The `add_outcome` method
would also fill in the relative fields based on the list of
key-value pairs available in the `values` argument.

This solution might have to be revisited once we start testing
more screens, especially because we should be moving the expecation
more to a more explicit place (from the `values` in the steps to
the feature/scenario itself).