Use a better sleep to set values
When a new value for Procurement Area and Access Point is entered,
a `submitForm` method is triggered from `OA_HTML/cabo/jsLibs/Common2_2_24_5.js`
and a "partial" submission of the form is basically made – this
causes a quick refresh of the page, but too slow to allow the
webdriver to find and fill the following elements of the page properly.
By overriding the `submitForm` method we can use it to mark whether
the form is reloading because of the partial submission and wait
before it's been reloaded completely.