Introduce CWA configuration loader library
A lot of knowledge is generally scattered across tests, fed with values
that are valuable but without a single source of truth to explain them.

For pricing, and maybe for validations too, most rules are "static" and
rarely change. It makes sense to summarise them in one place, load
as a configuration and focus on writing tests that know what the numbers
are. At the end of the day, these are _feature_ tests, so meant to quickly
test functionality, rather than describe all edge cases.

The SubmissionConfig library is a first attempt at isolating configuration,
and possibly re-use it in another context (non-testing related).