Rewrite Immigration and Asylum pricing feature tests
These tests use a different approach from the previous tests and are
meant to avoid making mistakes when trying out formulas. We also
want to avoid to write feature descriptions, but rather let the steps
print out what the conditions are for the tests.

A more natural language used by the business is introduced and input
is always made of a few words that map to logic embedded in the codebase.
Same for the expectations.

The embedding is minimal in the steps, which rely on the configuration
loaded from CWA.