Update pricing steps to take into account the escape fee threshold
A difference between categories of laws exists when looking at how the escape
fee threshold is used to flag cases that escape the standard fee (or the max price cap).

To accomodate the differences, the formula for the escape fee is now moved to
the configuration loader, while the implementation steps will simply rely on it.

The formula will be printed to screen and on the report when the adequate
assertion steps are used.

Finally, this commit also makes sure the max price cap is used when one exists.