Refactor Form field builder and Add Outcome page object
Form builder rewritten in an OO way.
It can now also return all fields and labels for an area and category of law.

The Add Outcame page object can now rely on those labels to create elements
dynamically in the DOM. This removes the need to keep this page up to date and
can help resolve conflicts between elements with a similar name.