refactor: refactored 1 existing step definition (sd) and replaced some tightly coupled sd's with a more flexible one
A decision was made to re-factor some sd's in cwa_outcome_steps to support a new feature.
This meant making an existing sd more flexible by allowing it to be called with an area of law
and category of law.  This info is used to retrieve the correct set of fields that need to be populated
together with any defaults.  This sd expects a data table from the feature tests.  It loops round
creating records based defaults/fields given the passed area of law/category of law and allowing the
data table to supply default value overrides.

In addition, there were some legacy sd's which were quite large and heavily tied to certain Legal Help
categories of law, as the name of the category of law was included in the sd name.  These have all
been replaced by a more flexible version.  This sd is called from feature tests using an example tables.
There is still some coupling as the more flexible replacement sd still expects a certain number of fixed fields.
But it can be passed an area of law and category of law (similar to the data table sd) and uses this to
grab the correct set of fields/defaults to use (overriding a fixed number of values).

For entering submissions into CWA from a feature test there are now 2 main approaches that can be taken.
Either the data table approach or the example table approach.  The sd's which implement them both make use
of the yaml files to retrieve fields/default values which are then overridden by values passed from the feature test.