New method in bulkload.rb to use ox xml lib and screen_field_builder to build and save a bulkload file based on info from the .yml field config files.  Slight re-org of some of the step defs in bulk_load_steps.rb so I could separate out the step which populates the search fields as this was tightly coupled with CWAProvider.  I couldn't find a way of only using CWAProvider to grab provider details without the code also trying to grab a list of category of law specific fields so I mocked up a MonthlySubmission method to provide the code with dummy info for the bulkload drafting process.
A WIP of course with ideas for how to move away from the current method of drafting the
bulkload file.  The ox xml lib was chosen as it claims to be much faster than the
current lib we are using and, for me at least, using it seemed very straightforward.