-bulk_load_steps.rb 4 new step definitions. 'the following error message is expected for each:' Takes an error message and asserts that this must be found on the bulkload errors page for each matter type.
'the following error message is matched for each:'
As above but uses the match operator to allow for regular expression matching.

'the following error messages are matched for each:'
Allows the tester to pass a table of error messages which are matched against the errors reported for each matter type on a bulk loaded outcome.

'there should be no error outcomes'
Simple check that there are no errors reported (problem, duplicate or invalid) and that the total no. of outcomes matches the total successful.

-fields/am_legal_help.yml
Add tolerance_indicator field.

-defaults/am_legal_help_defaults.yml
Add tolerance_indicator default.

-cwa_env.yml
Add providers for housing category of law together with value sets.

-hooks.rb
Support the creation of an html report and allow screenshots to be embedded when errors occur (if not running in headless mode).

To create a report make sure the html-report directory exists in the root and you pass in the following cucumber command line options (which can be embedded in the cucumber.yml):
--format html --out html-report/<report_filename>.html

-housing_validation_<1-7>.feature
Bulkload feature tests for housing category of law covering scope rule tests.