Remove redundant workflow filters from CI config
Background:

In CircleCI workflow config the `when` filter is used to control if a
workflow should be run in a pipeline or not. The default behaviour when
a `when` filter is not present is to always run the workflow.

The value for a workflow `when` filter should be a "logic statement"[1]
and if that statement evaluates to a truthy value then the workflow is
run.

The current logic statements are interpreted with a JavaScript-like
truthiness, particularly, a non-empty string is considered truthy.

There are no special constants in workflow `when` filters (unlike the
`when` attribute on a step), so a non-empty string constant is treated
as a truthy value, whatever the actual string contents.

In other words, `when: on_success` (YAML parses `on_success` as a
string) is equivalent to `when: true` (YAML parses `true` as a boolean),
which is equivalent to the default behaviour.

[1] https://circleci.com/docs/configuration-reference/#logic-statements